Moved iterator classes to inner classes
